chored(build): improve dependencies
- reverse core -> ledger dep
- remove ledger dep from p2p
