+++ core/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 license = "Apache-2.0"
 
 [dependencies]
-serde = "1.0.147"
+serde = { version = "1.0.147", features = ["rc"] }
 slab = { version = "0.4.7", features = ["serde"] }
 tracing = { version = "0.1", features = ["std"] }
 sha2 = "0.10.6"
@@ -14,4 +14,3 @@ tokio = { version = "1.26", features = ["sync"] }
 
 mina-hasher = { workspace = true }
 mina-p2p-messages = { workspace = true }
-ledger = { workspace = true }
