+++ core/src/snark/snark.rs
@@ -1,6 +1,5 @@
 use std::sync::Arc;
 
-use ledger::scan_state::scan_state::transaction_snark::work::Work;
 use mina_p2p_messages::binprot::{
     self,
     macros::{BinProtRead, BinProtWrite},
@@ -102,12 +101,3 @@ impl From<&Snark> for NetworkPoolSnarkPoolDiffVersionedStableV2AddSolvedWork1 {
     }
 }
 
-impl From<&Snark> for Work {
-    fn from(value: &Snark) -> Self {
-        Self {
-            prover: (&value.snarker).into(),
-            fee: (&value.fee).into(),
-            proofs: (&*value.proofs).into(),
-        }
-    }
-}
