+++ p2p/Cargo.toml
@@ -29,10 +29,8 @@ reqwest = { version = "0.11.22" }
 
 redux = { git = "https://github.com/openmina/redux-rs.git", branch="feat/global-time", features = ["serde"] }
 mina-p2p-messages = { workspace = true }
-ledger = { workspace = true }
 
 openmina-core = { path = "../core" }
-snark = { path = "../snark" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { version = "1.26", features = ["rt"] }
